backend/requirements.txt
google-genai>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
